requirements.txt
Pillow>=10.0.0
tkinter-dnd>=0.2.0
pyinstaller>=5.13.0